we embed ed25519 donna C code into a go file using cgo so there is no need to build it externally.
