+++ Makefile
@@ -131,12 +131,6 @@ crypto/libs/$(OS_TYPE)/$(ARCH)/lib/libsodium.a:
 		$(MAKE) && \
 		$(MAKE) install
 
-
-crypto/libs/$(OS_TYPE)/$(ARCH)/lib/ed25519-donna.o:
-	mkdir -p crypto/libs/$(OS_TYPE)/$(ARCH)/lib
-	gcc -Wall -std=c99 -DED25519_REFHASH -DED25519_CUSTOMRANDOM -Wno-macro-redefined -Icrypto/ed25519-donna -Wno-incompatible-pointer-types-discards-qualifiers crypto/ed25519-donna/ed25519.c -m64 -O3 -c -o crypto/libs/$(OS_TYPE)/$(ARCH)/lib/ed25519-donna.o
-
-
 deps:
 	./scripts/check_deps.sh
 
@@ -149,7 +143,7 @@ ALGOD_API_FILES := $(shell find daemon/algod/api/server/common daemon/algod/api/
 ALGOD_API_SWAGGER_INJECT := daemon/algod/api/server/lib/bundledSpecInject.go
 
 # Note that swagger.json requires the go-swagger dep.
-$(ALGOD_API_SWAGGER_SPEC): $(ALGOD_API_FILES) crypto/libs/$(OS_TYPE)/$(ARCH)/lib/libsodium.a crypto/libs/$(OS_TYPE)/$(ARCH)/lib/ed25519-donna.o
+$(ALGOD_API_SWAGGER_SPEC): $(ALGOD_API_FILES) crypto/libs/$(OS_TYPE)/$(ARCH)/lib/libsodium.a 
 	cd daemon/algod/api && \
 		PATH=$(GOPATH1)/bin:$$PATH \
 		go generate ./...
@@ -163,7 +157,7 @@ KMD_API_FILES := $(shell find daemon/kmd/api/ -type f | grep -v $(KMD_API_SWAGGE
 KMD_API_SWAGGER_WRAPPER := kmdSwaggerWrappers.go
 KMD_API_SWAGGER_INJECT := daemon/kmd/lib/kmdapi/bundledSpecInject.go
 
-$(KMD_API_SWAGGER_SPEC): $(KMD_API_FILES) crypto/libs/$(OS_TYPE)/$(ARCH)/lib/libsodium.a crypto/libs/$(OS_TYPE)/$(ARCH)/lib/ed25519-donna.o
+$(KMD_API_SWAGGER_SPEC): $(KMD_API_FILES) crypto/libs/$(OS_TYPE)/$(ARCH)/lib/libsodium.a 
 	cd daemon/kmd/lib/kmdapi && \
 		python3 genSwaggerWrappers.py $(KMD_API_SWAGGER_WRAPPER)
 	cd daemon/kmd && \
@@ -192,7 +186,7 @@ build: buildsrc gen
 # get around a bug in go build where it will fail
 # to cache binaries from time to time on empty NFS
 # dirs
-buildsrc: check-go-version crypto/libs/$(OS_TYPE)/$(ARCH)/lib/libsodium.a crypto/libs/$(OS_TYPE)/$(ARCH)/lib/ed25519-donna.o node_exporter NONGO_BIN deps $(ALGOD_API_SWAGGER_INJECT) $(KMD_API_SWAGGER_INJECT)
+buildsrc: check-go-version crypto/libs/$(OS_TYPE)/$(ARCH)/lib/libsodium.a  node_exporter NONGO_BIN deps $(ALGOD_API_SWAGGER_INJECT) $(KMD_API_SWAGGER_INJECT)
 	mkdir -p tmp/go-cache && \
 	touch tmp/go-cache/file.txt && \
 	GOCACHE=$(SRCPATH)/tmp/go-cache go install $(GOTRIMPATH) $(GOTAGS) $(GOBUILDMODE) -ldflags="$(GOLDFLAGS)" ./...
